1.完成首页的三个Fragment的网络请求和布局2.完成分类页面的布局和跳转3.完成个人中心页排版

+++ app/src/main/java/cn/xhl/client/manga/utils/SystemUtil.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.regex.Pattern;
 
 /**
- * Created by lixiuhao on 2017/5/5 0005.
+ * @author Mike on 2017/5/5 0005.
  */
 
 public class SystemUtil {
@@ -17,8 +17,8 @@ public class SystemUtil {
         return String.valueOf(System.currentTimeMillis() / 1000);
     }
 
-    public static long getTimestamp() {
-        return System.currentTimeMillis() / 1000;
+    public static int getTimestamp() {
+        return Long.valueOf(System.currentTimeMillis() / 1000).intValue();
     }
 
     /**
